Update buy-or-sell.mdx
`youtube: ba5Om3JoQp8`

###### Discover the difference

For sellers, Sotheby's International Realty Canada showcases your cottage or home to more quality buyers, locally and around the world. Our associates offer the highest calibre of service, discretion and marketing, regardless of your cottage or home's size, neighbourhood or price point. For buyers, Sotheby’s International Realty Canada provides skilled and authoritative assessment of real estate opportunities in cottage country, as well as exclusive access to some of the best cottage properties on the market many of which may never be available to the public on MLS.